refactor: set normal naming,rename function and etc.
